+++ odoo/songs/updates/update_10_1_1/sale_workflow_settings.py
@@ -5,23 +5,6 @@
 import anthem
 
 
-# @anthem.log
-# def create_journal_ir_property(ctx):
-#     automatic_workflow = ctx.env.ref(
-#         'sale_automatic_workflow.automatic_validation')
-#
-#     companies = ctx.env['res.company'].search([])
-#     admin_company = ctx.env.user.company_id
-#     for company in companies:
-#         ctx.env.user.company_id = company
-#         sale_journal = ctx.env['account.journal'].search([
-#             ('company_id', '=', company.id), ('type', '=', 'sale')])
-#         if sale_journal and len(sale_journal) == 1:
-#             automatic_workflow.property_journal_id = sale_journal.id
-#
-#     ctx.env.user.company_id = admin_company
-
-
 @anthem.log
 def configuration(ctx):
     automatic_workflow = ctx.env.ref(
@@ -46,4 +29,3 @@ def configuration(ctx):
 def main(ctx):
     """ Main: sale workflow settings """
     configuration(ctx)
-    # create_journal_ir_property(ctx)