+++ HISTORY.rst
@@ -11,6 +11,13 @@ Unreleased
 **Bugfixes**
 
 * DT-56: Fix deletion of partner linked to a company
+* DT-55: Fix sprint 4:
+
+ * A user can only see product from parents companies
+ * dt_accounting_group should not see agency customers
+ * Add widget selection on depiltech_payment_mode on sale_order form view
+ * Fix translation on sale discount program
+ * Configure sale automatic workflow settings
 
 **Build**